+++ packages/ai-constructs/src/conversation/runtime/event-tools-provider/event_tools_provider.ts
@@ -1,6 +1,7 @@
 import { ConversationTurnEvent, ExecutableTool } from '../types';
 import { GraphQlTool } from './graphql_tool';
 import { GraphQlQueryFactory } from './graphql_query_factory';
+import { UserAgentProvider } from '../user_agent_provider';
 
 /**
  * Creates executable tools from definitions in conversation turn event.
@@ -29,7 +30,7 @@ export class ConversationTurnEventToolsProvider {
         graphqlApiEndpoint,
         query,
         this.event.request.headers.authorization,
-        this.event.request.headers['x-amz-user-agent']
+        new UserAgentProvider(this.event)
       );
     });
     return tools ?? [];